BaseIdmResourceWithSearchTest class is introduced
