+++ src/test/java/gov/ca/cwds/idm/CreateUserTest.java
@@ -32,7 +32,7 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-public class CreateUserTest extends BaseIdmResourceTest {
+public class CreateUserTest extends BaseIdmResourceWithSearchTest {
 
   @Test
   @WithMockCustomUser(roles = {OFFICE_ADMIN}, adminOfficeIds = {"otherOfficeId"})
