+++ src/test/java/gov/ca/cwds/idm/UpdateUserTest.java
@@ -53,7 +53,7 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-public class UpdateUserTest extends BaseIdmResourceTest {
+public class UpdateUserTest extends BaseIdmResourceWithSearchTest {
 
   @Test
   @WithMockCustomUser
